requirements.txt
django-model-utils==3.1.2
webstack-django-jwt-auth==1.1.2
djangorestframework-jwt==1.11.0
djangorestframework==3.9.3
Django==2.2.1
requests==2.21.0
beautifulsoup4==4.7.1
pytz==2018.9
python-magic==0.4.15
django-cors-headers==3.0.0
pdf2image==1.5.4
Pillow==6.0.0
